pyproject.toml
[tool.poetry]
name = "kami-backend"
version = "0.1.0"
description = ""
authors = ["Arnaud Tsombeng <arnaud.tsombeng@gmail.com>"]
readme = "README.md"
packages = [{include = "kami"}]

[tool.poetry.dependencies]
python = "3.11.5"
djangorestframework = "3.15.2"
django = "5.1.4"
psycopg2 = "^2.9.10"
jinja2 = "^3.1.5"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.9.0"
factory-boy = "^3.3.1"
faker = "33.1.0"
jinja2 = "^3.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
